requirements.txt
python-telegram-bot
bs4
requests
setuptools
pyshorteners
validators
urlextract